fixed routes  as stated in swagger
