+++ app.js
@@ -6,7 +6,7 @@ require("dotenv").config();
 // const usersRouter = require("./routes/api/users");
 const transactionsRouter = require("./routes/api/transactions");
 const categoriesRouter = require("./routes/api/categories");
-const statisticsRouter = require("./routes/api/statistics");
+const statsRouter = require("./routes/api/stats");
 
 const app = express();
 
@@ -20,7 +20,8 @@ app.use(express.json());
 // app.use("/api/users", usersRouter);
 app.use("/api/transactions", transactionsRouter);
 app.use("/api/categories", categoriesRouter);
-app.use("/api/statistics", statisticsRouter);
+app.use("/api/transactions/stats", statsRouter);
+
 
 app.use((_, res) => {
   res.status(404).json({ message: "Not found" });
