--- controllers/statistics/index.js
@@ -1,5 +1,0 @@
-const getStatistics = require("./getStatistics");
-
-module.exports = {
-    getStatistics
-};
