+++ controllers/stats/index.js
@@ -1,0 +1,5 @@
+const getStats = require("./getStats");
+
+module.exports = {
+    getStats
+};
