--- routes/api/statistics.js
@@ -1,8 +1,0 @@
-const express = require("express")
-const router = express.Router()
-const { getStatistics } = require("../../controllers/statistics");
-
-router.get("/", getStatistics);
-
-module.exports = router;
-
