+++ routes/api/stats.js
@@ -1,0 +1,10 @@
+
+const express = require("express")
+const router = express.Router()
+const { getStats } = require("../../controllers/stats");
+
+router.get("/", getStats);
+
+module.exports = router;
+
+
