+++ routes/api/transactions.js
@@ -4,4 +4,5 @@ const transactionController = require("../../controllers/transaction");
 // const { auth } = require("../../middlewares/auth");
 router.post("/", transactionController.addTransactions);
 router.get("/", transactionController.getTransactions);
+
 module.exports = router;
